+++ eras/alonzo/impl/CHANGELOG.md
@@ -18,6 +18,7 @@
 
 ### `testlib`
 
+* Change type signature of `genPlutusScript`, `genNativeScript`, `genAlonzoScript`, `alwaysSucceedsLang` and `alwaysFailsLang`.
 * Remove `TxInfoLanguage` and `mkTxInfoLanguage`
 * Added `Era` module with `AlonzoEraTest` class